requirements.txt
alembic==1.10.2
anyio==3.6.2
bcrypt==4.0.1
black==23.3.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
Faker==17.6.0
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.6
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
